pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsreader"
version = "0.1.0"
description = "Newsreader application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psutil",
    "lxml[html_clean]",
    "newspaper3k==0.2.8",
    "nltk==3.8.1",
    "schedule==1.2.0",
    "bcrypt==4.0.1",
    "flask",
    "requests>=2.31.0",
    "spacy==3.7.2",
    "geopy==2.4.1"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
newsreader = ["templates/*.html"]

